- The Hey API client was improved to prevent the use of multiple instances. - A basic landing page was added. - The Auditor role was added for a user; this role does not have data modification permissions. - The properties of some tables were changed to make the data easier to understand. - Batches were added to the product view. - The container's post-install script has been improved.

+++ SigaBackend/Endpoints/CategoryEndpoints.cs
@@ -1,3 +1,4 @@
+using Microsoft.AspNetCore.Authorization;
 using SigaBackend.DTOs;
 using SigaBackend.Services;
 
@@ -9,18 +10,18 @@ public static class CategoryEndpoints
   {
     var group = routes.MapGroup("/api/categories");
 
-    group.MapPost("/", (CategoryCreateDto category, ICategoryService service) => service.CreateCategoryAsync(category)).WithName("CreateCategory").RequireAuthorization();
+    group.MapPost("/", (CategoryCreateDto category, ICategoryService service) => service.CreateCategoryAsync(category)).WithName("CreateCategory").RequireAuthorization(new AuthorizeAttribute { Roles = "Admin" });
 
-    group.MapGet("/", ([AsParameters] PaginationParams queryParams, ICategoryService service) => service.GetCategoriesAsync(queryParams)).WithName("GetCategories").RequireAuthorization();
+    group.MapGet("/", ([AsParameters] PaginationParams queryParams, ICategoryService service) => service.GetCategoriesAsync(queryParams)).WithName("GetCategories").RequireAuthorization(new AuthorizeAttribute { Roles = "Admin,Auditor" });
 
-    group.MapGet("/{id}", (int Id, ICategoryService service) => service.GetCategoryByIdAsync(Id)).WithName("GetCategoryById").RequireAuthorization();
+    group.MapGet("/{id}", (int Id, ICategoryService service) => service.GetCategoryByIdAsync(Id)).WithName("GetCategoryById").RequireAuthorization(new AuthorizeAttribute { Roles = "Admin,Auditor" });
 
-    group.MapGet("/{id}/products", (int Id, [AsParameters] PaginationParams queryParams, ICategoryService service) => service.GetProductsByCategoryAsync(Id, queryParams)).WithName("GetProductsByCategory").RequireAuthorization();
+    group.MapGet("/{id}/products", (int Id, [AsParameters] PaginationParams queryParams, ICategoryService service) => service.GetProductsByCategoryAsync(Id, queryParams)).WithName("GetProductsByCategory").RequireAuthorization(new AuthorizeAttribute { Roles = "Admin,Auditor" });
 
-    group.MapGet("/lookup", (ICategoryService service) => service.GetCategoriesLookupAsync()).WithName("GetCategoriesLookup").RequireAuthorization();
+    group.MapGet("/lookup", (ICategoryService service) => service.GetCategoriesLookupAsync()).WithName("GetCategoriesLookup").RequireAuthorization(new AuthorizeAttribute { Roles = "Admin,Auditor" });
 
-    group.MapPut("/{id}", (int Id, CategoryBasicDto dto, ICategoryService service) => service.UpdateCategoryAsync(Id, dto)).WithName("UpdateCategory").RequireAuthorization();
+    group.MapPut("/{id}", (int Id, CategoryBasicDto dto, ICategoryService service) => service.UpdateCategoryAsync(Id, dto)).WithName("UpdateCategory").RequireAuthorization(new AuthorizeAttribute { Roles = "Admin" });
 
-    group.MapDelete("/{id}", (int Id, ICategoryService service) => service.DeleteCategoryAsync(Id)).WithName("DeleteCategory").RequireAuthorization();
+    group.MapDelete("/{id}", (int Id, ICategoryService service) => service.DeleteCategoryAsync(Id)).WithName("DeleteCategory").RequireAuthorization(new AuthorizeAttribute { Roles = "Admin" });
   }
 }
